Make `cid()` work on binary blobs
Instead of having `util.cid()` take a DAG node as input, change it
to take a binary blob as input. This can save some (de)serializtion
overhead.

It also gains optional options for specifying a CID version and the
hash algorithm that should be used for calculation (if you don't
want to use the default one).