Rename properties
Renaming properties back to their original names.

defaultHashCode -> defaultHashAlg
codec - format

The intention was to make clear that the type is a multicodec. But
this shouldn't be done by the name of the property, but by a type
description.

Generally we should always use the codes and not the string identifiers
of the multicodecs.